feat: add inspector form to page before load
